Import CLV calculation
